minor touchup

+++ graphic_components/camera.hpp
@@ -8,8 +8,8 @@
 #include "SDL.h"
 
 struct Camera {
-    int x = 0;
-    int y = 0;
+    double x = 0;
+    double y = 0;
     float zoom = 1.f;
 };
